requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
PyMySQL
google-cloud-storage
python-multipart
Jinja2
email-validator
